server/pyproject.toml
[project]
name = "hms"
version = "0.1.0"
description = "Hospital Management System V2"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "flask>=3.1.3",
    "flask-sqlalchemy>=3.1",
    "flask-jwt-extended>=4.7",
    "flask-caching>=2.3.1",
    "argon2-cffi>=25.1",
    "celery[redis]>=5.4",
    "ruff>=0.15.9",
    "redis>=6.4.0",
]

[tool.ruff]
line-length = 100
